refactor OW config and prometheus patch
